Revert "timed graphic component for animation stuff"
This reverts commit 8369c0a27497381f9b6353348b05d0bd86f8bc5b.

+++ src/Timer.java
@@ -1,8 +1,8 @@
 public class Timer {
   private long startTime;
-  private double duration;
+  private long duration;
 
-  public Timer(double duration) {
+  public Timer(long duration) {
     // duration in seconds
     this.duration = duration*(1_000_000_000);
     this.startTime = System.nanoTime();
